+++ src/pages/Restrictions/components/MapLayers/MapLayers.tsx
@@ -9,8 +9,22 @@ import {
 } from '../../../../shared/map/mapLayers'
 
 import { useRestrictionsMapContext } from '../../contexts/MapContext'
-import { RestrictionsMapLayerVehicleAxleWeight } from './VehicleAxleWeight'
-import { RestrictionsMapLayerVehicleWidth } from './VehicleWidth'
+import {
+  vehicleAxleWeightLayerId,
+  vehicleHeightLayerId,
+  vehicleLengthLayerId,
+  vehicleWeightLayerId,
+  vehicleWidthLayerId,
+} from '../../contexts/mapLayersReducer'
+import {
+  axleWeightCategories,
+  heightCategories,
+  lengthCategories,
+  weightCategories,
+  widthCategories,
+} from '../../vehiclePropertyCategories'
+
+import { RestrictionsMapLayerVehicleProperty } from './VehicleProperty'
 
 function selectedBaseLayer(id: string) {
   return [aerialImages, topoBlackWhite].find(layer => layer.id === id)
@@ -21,9 +35,35 @@ export const RestrictionsMapLayers = () => {
 
   return (
     <>
-      <RestrictionsMapLayerVehicleAxleWeight />
+      <RestrictionsMapLayerVehicleProperty
+        categories={axleWeightCategories}
+        layerId={vehicleAxleWeightLayerId}
+        propertyName="axleWeight"
+      />
 
-      <RestrictionsMapLayerVehicleWidth />
+      <RestrictionsMapLayerVehicleProperty
+        categories={widthCategories}
+        layerId={vehicleWidthLayerId}
+        propertyName="width"
+      />
+
+      <RestrictionsMapLayerVehicleProperty
+        categories={heightCategories}
+        layerId={vehicleHeightLayerId}
+        propertyName="height"
+      />
+
+      <RestrictionsMapLayerVehicleProperty
+        categories={lengthCategories}
+        layerId={vehicleLengthLayerId}
+        propertyName="length"
+      />
+
+      <RestrictionsMapLayerVehicleProperty
+        categories={weightCategories}
+        layerId={vehicleWeightLayerId}
+        propertyName="weight"
+      />
 
       <TileLayer
         options={{ ...roadNetworkNoRestrictions.options, opacity: 0.5 }}