+++ src/pages/Restrictions/components/MapLegend/MapLegend.tsx
@@ -2,10 +2,25 @@ import { Column, Divider, RadioGroup, Row } from '@amsterdam/asc-ui'
 
 import { MapLegend } from '../../../../shared/components/MapLegend'
 
+import {
+  vehicleAxleWeightLayerId,
+  vehicleHeightLayerId,
+  vehicleLengthLayerId,
+  vehicleWeightLayerId,
+  vehicleWidthLayerId,
+} from '../../contexts/mapLayersReducer'
+
+import {
+  axleWeightCategories,
+  heightCategories,
+  lengthCategories,
+  weightCategories,
+  widthCategories,
+} from '../../vehiclePropertyCategories'
+
 import { RestrictionsMapLegendBaseLayers } from './BaseLayers'
 import { RestrictionsMapLegendRoadNetwork } from './RoadNetwork'
-import { RestrictionsMapLegendVehicleAxleWeight } from './VehicleAxleWeight'
-import { RestrictionsMapLegendVehicleWidth } from './VehicleWidth'
+import { RestrictionsMapLegendVehicleProperty } from './VehicleProperty'
 
 export const RestrictionsMapLegend = () => {
   return (
@@ -25,9 +40,35 @@ export const RestrictionsMapLegend = () => {
       <Row halign="flex-start" hasMargin={false}>
         <Column span={12}>
           <RadioGroup name="maplayers-vehicle-properties">
-            <RestrictionsMapLegendVehicleAxleWeight />
+            <RestrictionsMapLegendVehicleProperty
+              categories={axleWeightCategories}
+              layerId={vehicleAxleWeightLayerId}
+              label="Aslast"
+            />
 
-            <RestrictionsMapLegendVehicleWidth />
+            <RestrictionsMapLegendVehicleProperty
+              categories={widthCategories}
+              layerId={vehicleWidthLayerId}
+              label="Breedte"
+            />
+
+            <RestrictionsMapLegendVehicleProperty
+              categories={heightCategories}
+              layerId={vehicleHeightLayerId}
+              label="Hoogte"
+            />
+
+            <RestrictionsMapLegendVehicleProperty
+              categories={lengthCategories}
+              layerId={vehicleLengthLayerId}
+              label="Lengte"
+            />
+
+            <RestrictionsMapLegendVehicleProperty
+              categories={weightCategories}
+              layerId={vehicleWeightLayerId}
+              label="Totaal gewicht"
+            />
           </RadioGroup>
         </Column>
       </Row>